Backend Completed

+++ backend/controllers/contactController.js
@@ -2,7 +2,8 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getContacts = asyncHandler( async (req, res)=>{
-    const contacts = await Contact.find();
+    console.log(req.body);
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
@@ -13,7 +14,7 @@ const createContact = asyncHandler( async (req, res)=>{
         res.status(400);
         throw new Error("All fields are mandatory!");
     }
-    const contact = await Contact.create({name, email, phone});
+    const contact = await Contact.create({user_id: req.user.id, name, email, phone});
     res.status(201).json(contact);
 });
 
@@ -32,6 +33,10 @@ const updateContact = asyncHandler( async (req, res)=>{
         res.status(400);
         throw new Error("Contact not found");
     }
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update");
+    }
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body);
     res.status(200).json(updatedContact);
 });
@@ -39,8 +44,12 @@ const updateContact = asyncHandler( async (req, res)=>{
 const deleteContact = asyncHandler( async (req, res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
-        res.status(400);
+        res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
